Update the supporter database
We are now using the restforce gem to query the database by email.

The public API method is add_donation, but we are adding more methods
for now to build the functionality incrementally

We take the first created supporter when multiple records have
the same email

For now we are leaving some fields hardcoded. There are four
fields that are required to create a donation in Salesforce:

 - Amount: comes from the request
 - CloseDate: comes from the gateway, harcoded for now
 - Name: harcoded but is different for recurring donations
 - StageName: hardcoded but is different for recurring donations

These fields alone allow to create a one-off donation in Salesforce,
but for our application we need to ensure that the donation is
attached to a supporter.

The Salesforce API allows to create an opportunity without it being
associated to an account. The account is then empty. But we need
the account id to associate a donation with a supporter. So the
account id is a required field for the app.

For this app at the moment the only required attribute for supporters
is the email, since we use it to search a supporter. The Salesforce
required field is LastName.