Encapsulate the knowledge of donation-related data
This class extracts Salesforce and application required fields
out of the supporter database, and adds validations to them.

Calculation and validation (of field's values) seem like
different responsibilities. However, if we think in terms of
reasons to change, both seem to change for the same reasons,
i.e., it makes sense to add a validation everytime you add a new
field, which is why we decided to leave the two actions in the
same class for now.